Transfered VideoViewCount methods to a new class called Videos
