+++ channel_videos/Carla Zambelli.csv
@@ -1,3 +1,3 @@
 "Título","Número de visualizações"
-"GILBERTO CARVALHO PROCESSA CARLA ZAMBELLI E PERDE FEIO!","738"
-"O GILMAR HOJE É NOSSO ALIADO - DIZ DEPUTADO WADIH DOMOUS DO PT","802"
+"GILBERTO CARVALHO PROCESSA CARLA ZAMBELLI E PERDE FEIO!","815"
+"O GILMAR HOJE É NOSSO ALIADO - DIZ DEPUTADO WADIH DOMOUS DO PT","819"
