+++ channel_videos/Jair Bolsonaro.csv
@@ -1,6 +1,6 @@
 "Título","Número de visualizações"
-"Grande parte da imprensa & a calúnia: Enéas no campo de ideias está mais atual que nunca.","11373"
-"""Robôs"" em Maracanaú e Pacatuba. Obrigado meu Ceará.","45721"
-"REVISTA VEJA: Bolsonaro usa robôs de última geração para obter sucesso na internet. kkk","106557"
-"BOLSONARO: ENTREVISTA AO PROGRAMA DIRETO DE BRASÍLIA.","82632"
-"“MÍDIA ESCULACHADA”. (Via Canal Hipócritas).","23588"
+"Grande parte da imprensa & a calúnia: Enéas no campo de ideias está mais atual que nunca.","12240"
+"""Robôs"" em Maracanaú e Pacatuba. Obrigado meu Ceará.","46386"
+"REVISTA VEJA: Bolsonaro usa robôs de última geração para obter sucesso na internet. kkk","108666"
+"BOLSONARO: ENTREVISTA AO PROGRAMA DIRETO DE BRASÍLIA.","85751"
+"“MÍDIA ESCULACHADA”. (Via Canal Hipócritas).","23636"
