+++ channel_videos/MAB comunicação.csv
@@ -1,5 +1,5 @@
 "Título","Número de visualizações"
-"Arpilleras recebe premio de melhor documentário do SESC!","34"
+"Arpilleras recebe premio de melhor documentário do SESC!","35"
 "A água em disputa (FAMA 2018)","306"
 "A Saga dos Atingidos por Barragens! (parte 04)","37"
 "A Saga dos Atingidos por Barragens! (Parte 03)","24"
