+++ channel_videos/MTST Movimento dos Trabalhadores Sem Teto.csv
@@ -1,3 +1,3 @@
 "Título","Número de visualizações"
-"SOBRE A CONDENAÇÃO DE LULA E OS DESAFIOS DA LUTA","351"
+"SOBRE A CONDENAÇÃO DE LULA E OS DESAFIOS DA LUTA","352"
 "DIA 24 DE JANEIRO VAMOS TOMAR AS RUAS PELA DEMOCRACIA","188"
