+++ channel_videos/Michel Temer.csv
@@ -1,5 +1,5 @@
 "Título","Número de visualizações"
-"Presidente Michel Temer sanciona projeto de lei para segurança pública","27"
+"Presidente Michel Temer sanciona projeto de lei para segurança pública","29"
 "Presidente Michel Temer defende união para manter conquistas no País","95"
 "Diplomacia fortalece a voz do Brasil no mundo, destaca presidente Michel Temer","52"
 "Decreto permitirá estudos sobre modelo de capitalização da Eletrobras","36"
