+++ channel_videos/Movimento Agora.csv
@@ -1,6 +1,6 @@
 "Título","Número de visualizações"
-"Quem pode mudar a política no Brasil?","20"
+"Quem pode mudar a política no Brasil?","32"
 "Pesquisa Saúde Agora!","3"
 "Dia do índio - Anápuaká Tupinambá","5"
-"Semanas Temáticas - Agora!","5"
+"Semanas Temáticas - Agora!","6"
 "Conheça os membros do Agora! - Inês Lafer","12"
