+++ channel_videos/Movimento Brasil Livre.csv
@@ -1,6 +1,6 @@
 "Título","Número de visualizações"
-"Dirceu perde a esperança de sair da cadeia","832"
-"Em vídeo patético, Lula mostra um Brasil que ficou para trás!","9102"
-"SINDICATOS CONSEGUEM LIMINAR PARA COBRAR IMPOSTO SINDICAL","5640"
-"Projeto de lei 188/2018 - Pichadores terão de pagar multa e reparar vandalismo","8587"
-"Teste do Sofá ep. 28 | Nicholas Vital","9051"
+"UGANDA CONECTION /  PRESISDENTE DE UGANDA PROIBE SEXO ORAL, MC CAROL, CAMILA PITANGA DIZ.","196"
+"MBL NEWS | 23/04/18 | Paraguai conservador, delírio de Lula e frustrações no PSDB","3424"
+"Dirceu perde a esperança de sair da cadeia","2602"
+"Em vídeo patético, Lula mostra um Brasil que ficou para trás!","12318"
+"SINDICATOS CONSEGUEM LIMINAR PARA COBRAR IMPOSTO SINDICAL","6093"
