+++ channel_videos/Mídia NINJA.csv
@@ -1,6 +1,6 @@
 "Título","Número de visualizações"
-"13 dias de resistência em Curitiba - Acampamento Lula Livre","1541"
-"A Casa da Democracia é uma realidade!","1231"
-"12 dias de ocupação em Curitiba - Apoio internacional à Lula","4107"
-"Adolfo Pérez Esquivel indica Lula ao Prêmio Nobel da Paz","2672"
-"Lindbergh Farias - Desmilitarização da Polícia Militar","1685"
+"13 dias de resistência em Curitiba - Acampamento Lula Livre","1556"
+"A Casa da Democracia é uma realidade!","1250"
+"12 dias de ocupação em Curitiba - Apoio internacional à Lula","4140"
+"Adolfo Pérez Esquivel indica Lula ao Prêmio Nobel da Paz","2681"
+"Lindbergh Farias - Desmilitarização da Polícia Militar","1692"
