+++ channel_videos/Revoltados ON LINE.csv
@@ -1,6 +1,6 @@
 "Título","Número de visualizações"
-"Eleições 2018 sem VOTO IMPRESSO é GOLPE com Felipe Gimenez -Nilton Cacaos  | (Marcello Reis)","145"
-"Gleisi Hoffmann quando sair da prisão já tem proposta de emprego de Oscar Maroni (Marcello Reis)","543"
-"Gleisi Hoffmann convoca terroristas pela TV Al Jazeera - Cadeia nela (Marcello Reis)","730"
-"Bandido Lula liga da cadeia para Silvio Santos | (Marcello Reis)","917"
-"Jair Bolsonaro dá os parabéns a Marco Antonio Villa | (Marcello Reis)","2551"
+"Eleições 2018 sem VOTO IMPRESSO é GOLPE com Felipe Gimenez -Nilton Cacaos  | (Marcello Reis)","178"
+"Gleisi Hoffmann quando sair da prisão já tem proposta de emprego de Oscar Maroni (Marcello Reis)","584"
+"Gleisi Hoffmann convoca terroristas pela TV Al Jazeera - Cadeia nela (Marcello Reis)","732"
+"Bandido Lula liga da cadeia para Silvio Santos | (Marcello Reis)","918"
+"Jair Bolsonaro dá os parabéns a Marco Antonio Villa | (Marcello Reis)","2636"
