+++ channel_videos/TV CONTAG.csv
@@ -1,6 +1,6 @@
 "Título","Número de visualizações"
 "POLÍTICA DE COTA PARA JUVENTUDE RURAL","9"
-"ACAMPAMENTO ""LULA LIVRE"" - BRASÍLIA/DF","138"
+"ACAMPAMENTO ""LULA LIVRE"" - BRASÍLIA/DF","140"
 "Jamais poderão aprisionar nossos sonhos!","87"
 "Os povos do cerrado estão resistindo!","76"
 "Precisamos de ÁGUA para alimentar o Brasil!","194"
