+++ channel_videos/Vem Pra Rua Brasil.csv
@@ -1,6 +1,6 @@
 "Título","Número de visualizações"
 "Desembargadores do TRF-4 e seus familiares estão sofrendo ameaças. Isso é inaceitável!","537"
-"Lula persegue a Justiça enquanto reclama de perseguição","227"
+"Lula persegue a Justiça enquanto reclama de perseguição","228"
 "Nelson Nahim, condenado por estupro em primeira instância, assumirá vaga de deputado na Câmara","216"
 "Lula botou a culpa de tudo na Marisa.","152"
 "Desembargador Victor Laus, não ceda às pressões por adiamento","382"
