+++ channel_videos/ubesoficial.csv
@@ -1,5 +1,5 @@
 "Título","Número de visualizações"
-"(CHAMADA) Coração de Estudante - Pela Democracia e Liberdade do Lula","62"
+"(CHAMADA) Coração de Estudante - Pela Democracia e Liberdade do Lula","64"
 "Jornada de Lutas Edson Luís","66"
-"Convocação para a Jornada de Lutas Edson Luís","70"
+"Convocação para a Jornada de Lutas Edson Luís","71"
 "Samba da Intervenção","106"
