+++ channel_videos/uneoficial.csv
@@ -1,6 +1,6 @@
 "Título","Número de visualizações"
-"Jornada de Lutas da UBES - Edson Luís e Marielle Franco, presentes!","28"
-"Boletim Nossa Voz 23/03/18","283"
+"Jornada de Lutas da UBES - Edson Luís e Marielle Franco, presentes!","29"
+"Boletim Nossa Voz 23/03/18","296"
 "Boletim Nossa Voz 16 03 2018","46"
 "Vem aí... UNE Volante!","140"
 "Boletim Nossa Voz 09/03/2018","46"
