+++ data/youtube.csv
@@ -2,64 +2,64 @@
 "Frente Brasil Popular","","UCX2Aanu4fGewmhP4rf5GQ3Q","6","1851","288"
 "Frente Povo Sem Medo","","","","",""
 "Ocupa Política","","","","",""
-"Movimento Agora!","","UC5AH5khzIk15T0waSp-ayCg","96","6954","161"
+"Movimento Agora!","","UC5AH5khzIk15T0waSp-ayCg","96","6961","161"
 "Movimento Muitxs","","UCVoyVm_HowbEYG7Wqptx5eA","4","3720","91"
 "Bancada Ativista","","","","",""
 "RenovaBR","","","","",""
 "Frente Nacional Contra a censura","","","","",""
 "Frente Autônoma LGBT","","","","",""
 "Frente Nacional pela legalização do aborto","","","","",""
-"Mídia Ninja","midianinjafly","UCwefw3viiiaxprsV6jgJvEA","409","2107184","46480"
+"Mídia Ninja","midianinjafly","UCwefw3viiiaxprsV6jgJvEA","409","2107500","46493"
 "342 Artes","","","","",""
-"MTST - Movimento Trabalhadores Sem-Teto","","UC3OzrZMhnmEgVtxpJoDRkeg","34","22234","763"
+"MTST - Movimento Trabalhadores Sem-Teto","","UC3OzrZMhnmEgVtxpJoDRkeg","34","22235","763"
 "MST - Movimento dos Trabalhadores Sem Terra","","","","",""
 "MPL - Movimento Passe Livre","","","","",""
 "MLB - Movimento de Luta nos Bairros, Vilas e Favelas","","","","",""
-"UNE - União Nacional dos Estudantes","uneoficial","UCI9UhLCYUB8v82jdVvJMnuA","427","315935","2915"
+"UNE - União Nacional dos Estudantes","uneoficial","UCI9UhLCYUB8v82jdVvJMnuA","427","315945","2916"
 "CUT - Central Única dos Trabalhadores","","","","",""
-"Levante Popular da Juventude","","UCmo_qwwtofscJjjND_Pm-UA","156","212370","2669"
-"Juntos","jornaljuntos","UCmZp0eVq7fpxKHHvncJHSrg","242","195792","953"
+"Levante Popular da Juventude","","UCmo_qwwtofscJjjND_Pm-UA","156","212370","2670"
+"Juntos","jornaljuntos","UCmZp0eVq7fpxKHHvncJHSrg","242","195794","953"
 "MAIS","","UCXrqijOy0tezQ8_-0OCqAPw","41","12354","257"
 "Movimento Correnteza","","UCH2R0pAFyqnCtp253XVYq-w","1","1158","73"
-"UJS - União da Juventude Socialista","","UCz0799M83kYvb6qBkHxTWFQ","56","75901","1297"
+"UJS - União da Juventude Socialista","","UCz0799M83kYvb6qBkHxTWFQ","56","75910","1297"
 "UJR Brasil","","","","",""
 "Movimento de Mulheres Olga Benário - Nacional","","","","",""
 "Faísca - Anticapitalista e Revolucionária","","","","",""
 "Pão e Rosas Brasil","","UCDyMmoCj1jvF8ooGs_SQbJg","7","278","2"
 "MRT - Movimento Revolucionário dos Trabalhadores","","","","",""
 "UBES - União Brasileira Estudantes Secundaristas","UBESCOMUNICA","UCQ7QEpwjFy293Ir3I7Is_Cw","140","55869","1056"
-"União da Juventude Comunista - UJC Brasil","","UCLTOw6dtkalzyBRXvUuEpTg","188","67238","1723"
+"União da Juventude Comunista - UJC Brasil","","UCLTOw6dtkalzyBRXvUuEpTg","188","67238","1724"
 "RUA - Juventude Anticapitalista","","UCTSv82hMEIFnJjCOHzizVzw","4","7333","95"
 "Afronte","","","","",""
 "União Brasileira de Mulheres UBM","","","","",""
-"Movimento dos Atingidos por Barragens (MAB)","mabcomunicacao","UC3UDbLgGfCfxiHDc7m8faQQ","192","214600","1030"
+"Movimento dos Atingidos por Barragens (MAB)","mabcomunicacao","UC3UDbLgGfCfxiHDc7m8faQQ","192","214603","1030"
 "Contag","ContagBrasil","UCfi7PCkseJNwGbaD0X-mAXA","193","73817","586"
 "Geledés Mulher Negra","","UCO9-aYrQfWP1Mm0A0hJdbMA","1","275","147"
-"Marcha das Mulheres","mulheresemmarcha","UCh0Eth5qo5-S1IMigJhTLdg","120","311327","805"
-"Não Me Khalo","","UCLMp2AgzSZrEZhz1XzJXeXw","30","156568","3072"
-"Vem Pra Rua","","UCNORbWRHIwZuCiHz2Ukh6Nw","80","73353","1906"
-"Revoltados Online","revoltadosonline","UCVo8EvHIDfSaBQT3gSFUyxQ","859","7560900","42046"
-"Estudantes Pela Liberdade","libertarianismoplus","UCWU5PdQ9Ec5eZleRm6Qd-Ig","778","5254830","44427"
-"Escola Sem Partido","","UCTLIxRJaYAtf44baTh0MN6A","5","92069","682"
+"Marcha das Mulheres","mulheresemmarcha","UCh0Eth5qo5-S1IMigJhTLdg","120","311337","805"
+"Não Me Khalo","","UCLMp2AgzSZrEZhz1XzJXeXw","30","156577","3073"
+"Vem Pra Rua","","UCNORbWRHIwZuCiHz2Ukh6Nw","80","73354","1906"
+"Revoltados Online","revoltadosonline","UCVo8EvHIDfSaBQT3gSFUyxQ","859","7561112","42049"
+"Estudantes Pela Liberdade","libertarianismoplus","UCWU5PdQ9Ec5eZleRm6Qd-Ig","778","5254872","44429"
+"Escola Sem Partido","","UCTLIxRJaYAtf44baTh0MN6A","5","92088","681"
 "Instituto Mercado Popular","","","","",""
 "Libertarianismo +","","","","",""
-"Movimento Brasil Livre","","UC8QAdpiEWAOg3AOCCFDCOYw","1108","33694992","294143"
-"Movimento Contra a Corrupção","canalcontracorrupcao","UCvo3PC2zBfG1IsCWVJVn6pw","70","1663326","26974"
+"Movimento Brasil Livre","","UC8QAdpiEWAOg3AOCCFDCOYw","1111","33721184","294296"
+"Movimento Contra a Corrupção","canalcontracorrupcao","UCvo3PC2zBfG1IsCWVJVn6pw","70","1663347","26975"
 "Livres","","","","",""
-"Nas Ruas","nasruastv","UCX9VzlP1DyeasSivEr9Eddw","695","2205666","19660"
+"Nas Ruas","nasruastv","UCX9VzlP1DyeasSivEr9Eddw","695","2206055","19661"
 "Lula","","","","",""
 "Fernando Haddad","","","","",""
-"Ciro Gomes","","UCz0i4BaFbRuLun82sl5yWLw","132","1528881","17870"
+"Ciro Gomes","","UCz0i4BaFbRuLun82sl5yWLw","132","1528993","17871"
 "Dória","","","","",""
 "Alckmin","","","","",""
-"Bolsonaro","jbolsonaro","UC8hGUtfEgvvnp6IaHSAg1OQ","740","36813984","513463"
+"Bolsonaro","jbolsonaro","UC8hGUtfEgvvnp6IaHSAg1OQ","740","36821923","513701"
 "Henrique Meirelles","","","","",""
-"Temer","MichelTemer","UCI1scpqZalSY96yDKAp0afQ","977","417924","0"
-"Marina Silva","msilvaonline","UC9uefWa6TXIPDRWGZYMcTuA","872","4274055","13180"
-"Levy Fidelix","","UCHn-xluSQGIiknQ56iNTNdA","344","222353","2828"
-"Everaldo","BlogdoEveraldo","UClHDtWJ0HwBMD6j36C24w9A","137","170028","491"
+"Temer","MichelTemer","UCI1scpqZalSY96yDKAp0afQ","977","417945","0"
+"Marina Silva","msilvaonline","UC9uefWa6TXIPDRWGZYMcTuA","872","4274094","13183"
+"Levy Fidelix","","UCHn-xluSQGIiknQ56iNTNdA","344","222362","2830"
+"Everaldo","BlogdoEveraldo","UClHDtWJ0HwBMD6j36C24w9A","137","170029","491"
 "Flávio Rocha","","","","",""
-"Roberto Justus Ofici","robertoljustus","UCLWVlQxY7Dc4OFhc08Qjf9A","23","418924","1869"
+"Roberto Justus Ofici","robertoljustus","UCLWVlQxY7Dc4OFhc08Qjf9A","23","418924","1870"
 "Guilherme Boulos","","","","",""
 "Jaques Wagner","","","","",""
 "Fernando Collor de Mello","","","","",""
